+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## [0.1.1](https://github.com/sheldonWan/picture2color/compare/v0.1.0...v0.1.1) (2020-12-23)
+
+
+### Bug Fixes
+
+* fixed count bug ([2b1a699](https://github.com/sheldonWan/picture2color/commit/2b1a69969c5b774c8a301cd6bf2a159da31f8287))
+
+
+
+# [0.1.0](https://github.com/sheldonWan/picture2color/compare/v0.0.10...v0.1.0) (2020-12-23)
+
+
+### Bug Fixes
+
+* hex string bug fixed ([152ae7c](https://github.com/sheldonWan/picture2color/commit/152ae7cc80bb28a63d12248fa0263e8bc28ccd8e))
+
+
+### Features
+
+* 添加types和paramsFilter ([1540475](https://github.com/sheldonWan/picture2color/commit/1540475e64fbf679e28c6c7165a1d224e2b796f9))
+* 重新规划0.1.0版本 ([2c7861f](https://github.com/sheldonWan/picture2color/commit/2c7861f87e89baec82dc5e189ddc4f581967488c))
+* 重新设计color proxy ([7581427](https://github.com/sheldonWan/picture2color/commit/758142795585a103984f94dbfcc82e76a485ad54))
+
+
+
 ## [0.0.10](https://github.com/sheldonWan/picture2color/compare/v0.0.9...v0.0.10) (2020-11-27)
 
 
@@ -27,6 +52,10 @@
 
 
 ## [0.0.4](https://github.com/sheldonWan/picture2color/compare/0.0.3...0.0.4) (2020-10-22)
+
+
+
+## 0.0.3 (2020-10-22)